Add tab and modify UI type

+++ src/_01_ProductTest/DeleteProduct.java
@@ -1,6 +1,10 @@
 package _01_ProductTest;
 
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.ArrayList;
+import java.util.Collection;
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -8,27 +12,48 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.google.gson.Gson;
+
+import _14_Member.ResultBean;
 
 
-@WebServlet("/_01_ProductTest/DeleteServelet")
+
+@WebServlet("/_01_ProductTest/DeleteServlet")
 public class DeleteProduct extends HttpServlet {
 	private static final long serialVersionUID = 1L;
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		doPost(request, response);
+	}
     public void doPost(HttpServletRequest request,HttpServletResponse response)
             throws IOException, ServletException {
+    
+    	request.setCharacterEncoding("UTF-8");
+		response.setContentType("application/json; charset=UTF-8");
     	
-    	String productIdStr = request.getParameter("productId");
-    	boolean success = new ProductDAO().delete(productIdStr);
-    	System.out.println("刪除產品資料: "+ productIdStr);
+		String productIdStr = request.getParameter("productId");   	   	
+		Collection<String> coll = new ArrayList<>();
     	
-    	request.setAttribute("productId", productIdStr);
-    	
-    	if(success == true){
-    		RequestDispatcher rd = request.getRequestDispatcher("DeleteSuccess.jsp");
-    		rd.forward(request, response);
-    	}else{
-    		RequestDispatcher rd = request.getRequestDispatcher("DeleteError.jsp");
-    		rd.forward(request, response);		
-    	}
-    }
-    	
-}   
+    	if (productIdStr == null || productIdStr.trim().length() == 0) {
+			coll.add("欄位必須輸入序號");
+	
+		} else {
+			boolean success = new ProductDAO().delete(productIdStr);
+			System.out.println("刪除產品序號"+ productIdStr +"資料");
+
+			if (success == true) {
+				coll.add(("資料 "+productIdStr+"刪除成功"));
+				
+			} else {
+				coll.add(("資料 "+productIdStr+"刪除失敗"));
+			}
+		}
+		try (PrintWriter out = response.getWriter();) {
+
+			String toJson = new Gson().toJson(coll);
+			System.out.println(toJson);
+			out.println(toJson);
+		}
+
+	}
+}
